scripts/dock_example.py
#!/usr/bin/env python3
"""
Example protein-ligand docking workflow for Mars-adapted enzymes
Dependencies: rdkit, biopython, numpy

This module demonstrates molecular docking simulations for enzymes
adapted to Mars atmospheric conditions, focusing on CO2 fixation
and extremophile metabolic pathways.
"""

import argparse
import logging
import sys
from pathlib import Path
from typing import List, Tuple, Optional

try:
    import numpy as np
    from rdkit import Chem
    from rdkit.Chem import Descriptors, rdMolDescriptors
    from rdkit.Chem import AllChem
except ImportError as e:
    print(f"Missing required dependencies: {e}")
    print("Install with: pip install rdkit numpy")
    sys.exit(1)

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)


class MarsEnzymeDocking:
    """
    Molecular docking simulation for Mars-adapted enzymes
    
    Simulates protein-ligand interactions under Mars environmental
    conditions including low temperature, high radiation, and
    CO2-rich atmosphere.
    """
    
    def __init__(self, temperature: float = 233.15, pressure: float = 0.006):
        """
        Initialize docking simulation parameters
        
        Args:
            temperature: Mars surface temperature in Kelvin (default: -40°C)
            pressure: Mars atmospheric pressure in atm (default: 0.6% of Earth)
        """
        self.temperature = temperature
        self.pressure = pressure
        self.mars_conditions = {
            'temperature': temperature,
            'pressure': pressure,
            'co2_concentration': 0.96,  # 96% CO2 in Mars atmosphere
            'radiation_level': 'high'
        }
        logger.info(f"Initialized Mars docking conditions: T={temperature}K, P={pressure}atm")
    
    def generate_mars_ligands(self) -> List[Chem.Mol]:
        """
        Generate small molecule ligands relevant to Mars chemistry
        
        Returns:
            List of RDKit molecule objects for Mars-relevant compounds
        """
        # Mars-relevant SMILES strings
        mars_compounds = [
            'O=C=O',           # CO2 - primary atmospheric component
            'O',               # H2O - water ice
            'N#N',             # N2 - atmospheric nitrogen
            'O=O',             # O2 - trace atmospheric oxygen
            '[CH4]',           # CH4 - methane (detected in atmosphere)
            'C(=O)O',          # Formic acid - potential metabolite
            'CC(=O)O',         # Acetic acid - organic compound
            'C(C(=O)O)O',      # Glycolic acid - potential Mars organic
        ]
        
        molecules = []
        for smiles in mars_compounds:
            try:
                mol = Chem.MolFromSmiles(smiles)
                if mol is not None:
                    # Add hydrogens and generate 3D coordinates
                    mol = Chem.AddHs(mol)
                    AllChem.EmbedMolecule(mol, randomSeed=42)
                    molecules.append(mol)
                    logger.debug(f"Generated ligand: {smiles}")
            except Exception as e:
                logger.warning(f"Failed to generate molecule from {smiles}: {e}")
        
        logger.info(f"Generated {len(molecules)} Mars-relevant ligands")
        return molecules
    
    def calculate_mars_binding_affinity(self, ligand: Chem.Mol) -> float:
        """
        Calculate binding affinity adjusted for Mars conditions
        
        Args:
            ligand: RDKit molecule object
            
        Returns:
            Estimated binding affinity in kcal/mol (more negative = stronger binding)
        """
        # Calculate molecular descriptors
        mw = Descriptors.MolWt(ligand)
        logp = Descriptors.MolLogP(ligand)
        hbd = rdMolDescriptors.CalcNumHBD(ligand)
        hba = rdMolDescriptors.CalcNumHBA(ligand)
        
        # Simple scoring function adjusted for Mars conditions
        # Lower temperature increases binding affinity
        temp_factor = 298.15 / self.temperature  # Reference to room temp
        
        # Pressure effects on binding
        pressure_factor = 1.0 + (1.0 - self.pressure) * 0.1
        
        # Basic binding affinity estimation
        base_affinity = -1.5 * logp - 0.1 * mw + 0.5 * (hbd + hba)
        
        # Apply Mars condition corrections
        mars_affinity = base_affinity * temp_factor * pressure_factor
        
        return mars_affinity
    
    def dock_mars_enzyme_substrate(self, enzyme_name: str = "carbonic_anhydrase") -> dict:
        """
        Demonstrate docking workflow for Mars atmospheric processing
        
        Args:
            enzyme_name: Name of the target enzyme for docking
            
        Returns:
            Dictionary containing docking results and analysis
        """
        logger.info(f"Starting docking simulation for {enzyme_name}")
        
        # Generate Mars-relevant ligands
        ligands = self.generate_mars_ligands()
        
        if not ligands:
            raise ValueError("No valid ligands generated for docking")
        
        # Simulate docking results
        results = {
            'enzyme': enzyme_name,
            'conditions': self.mars_conditions,
            'ligand_results': []
        }
        
        for i, ligand in enumerate(ligands):
            try:
                # Calculate binding affinity
                affinity = self.calculate_mars_binding_affinity(ligand)
                
                # Get molecular properties
                smiles = Chem.MolToSmiles(ligand)
                mw = Descriptors.MolWt(ligand)
                
                ligand_result = {
                    'ligand_id': i,
                    'smiles': smiles,
                    'molecular_weight': round(mw, 2),
                    'binding_affinity': round(affinity, 3),
                    'mars_adapted_score': round(affinity * temp_factor, 3)
                }
                
                results['ligand_results'].append(ligand_result)
                logger.debug(f"Docked ligand {i}: {smiles} (ΔG = {affinity:.3f} kcal/mol)")
                
            except Exception as e:
                logger.error(f"Failed to dock ligand {i}: {e}")
        
        # Sort by binding affinity (most negative = best)
        results['ligand_results'].sort(key=lambda x: x['binding_affinity'])
        
        logger.info(f"Docking complete: {len(results['ligand_results'])} successful results")
        return results
    
    def print_docking_summary(self, results: dict) -> None:
        """Print a formatted summary of docking results"""
        print(f"\n🧬 Mars Enzyme Docking Results for {results['enzyme']}")
        print("=" * 60)
        print(f"Conditions: T={self.temperature}K, P={self.pressure}atm")
        print(f"CO2 concentration: {self.mars_conditions['co2_concentration']*100}%")
        print("\nTop Binding Candidates:")
        print("-" * 60)
        
        for i, ligand in enumerate(results['ligand_results'][:5]):  # Top 5
            print(f"{i+1:2d}. {ligand['smiles']:15s} "
                  f"ΔG: {ligand['binding_affinity']:6.2f} kcal/mol "
                  f"MW: {ligand['molecular_weight']:6.1f}")
        
        if len(results['ligand_results']) > 5:
            print(f"... and {len(results['ligand_results']) - 5} more candidates")


def main():
    """Main function for command-line usage"""
    parser = argparse.ArgumentParser(
        description="Mars enzyme docking simulation",
        formatter_class=argparse.RawDescriptionHelpFormatter,
        epilog="""
Examples:
  python dock_example.py
  python dock_example.py --enzyme carbonic_anhydrase --temp 223
  python dock_example.py --verbose
        """
    )
    
    parser.add_argument(
        '--enzyme', 
        default='carbonic_anhydrase',
        help='Target enzyme for docking (default: carbonic_anhydrase)'
    )
    parser.add_argument(
        '--temp', 
        type=float, 
        default=233.15,
        help='Mars surface temperature in Kelvin (default: 233.15)'
    )
    parser.add_argument(
        '--pressure', 
        type=float, 
        default=0.006,
        help='Mars atmospheric pressure in atm (default: 0.006)'
    )
    parser.add_argument(
        '--verbose', 
        action='store_true',
        help='Enable verbose logging'
    )
    
    args = parser.parse_args()
    
    if args.verbose:
        logging.getLogger().setLevel(logging.DEBUG)
    
    try:
        # Initialize docking simulation
        docker = MarsEnzymeDocking(temperature=args.temp, pressure=args.pressure)
        
        # Run docking simulation
        results = docker.dock_mars_enzyme_substrate(enzyme_name=args.enzyme)
        
        # Display results
        docker.print_docking_summary(results)
        
        # Save results to file
        output_file = Path(f"docking_results_{args.enzyme}.txt")
        with open(output_file, 'w') as f:
            f.write(f"Mars Enzyme Docking Results\n")
            f.write(f"Enzyme: {results['enzyme']}\n")
            f.write(f"Conditions: {results['conditions']}\n")
            f.write(f"Results: {len(results['ligand_results'])} ligands\n\n")
            
            for ligand in results['ligand_results']:
                f.write(f"{ligand['smiles']}\t{ligand['binding_affinity']:.3f}\t{ligand['molecular_weight']:.1f}\n")
        
        print(f"\n📁 Results saved to: {output_file}")
        print("✅ Docking simulation completed successfully!")
        
    except Exception as e:
        logger.error(f"Docking simulation failed: {e}")
        sys.exit(1)


if __name__ == "__main__":
    main()
