requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Computational biology
biopython>=1.79
rdkit-pypi>=2022.3.5
cobra>=0.26.0

# Development dependencies  
pytest>=6.2.0
flake8>=3.9.0
black>=21.6.0
mypy>=0.910
pytest-cov>=2.12.0

# Additional scientific computing
scikit-learn>=1.0.0
networkx>=2.6.0
plotly>=5.0.0

# File format support
lxml>=4.6.0
xmlschema>=1.8.0
openpyxl>=3.0.0
